requirements.txt
django==3.2.19
djangorestframework==3.14.0
